pyproject.toml
[project]
name = "uzbek-stt-bot"
version = "0.1.0"
description = "Phase 1 Telegram bot prototype for Uzbek speech evaluation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26.0,<2.0.0",
  "python-telegram-bot>=22.0",
  "python-dotenv>=1.0.1",
  "av>=14.0.0",
  "torch>=2.2.0,<2.9.0",
  "transformers>=4.40.0,<6.0.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
modal = [
  "modal>=0.73.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel>=0.43.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
